V54

+++ app/Http/Controllers/Admin/PostsController.php
@@ -14,7 +14,10 @@ class PostsController extends Controller
 {
 	public function index()
 	{
-		$posts = Post::all();
+		$posts = Post::allowed()->get();
+		// $posts = Post::all();
+		// $posts = Post::where('user_id', auth()->id() )->get();
+		// $posts = auth()->user()->posts;
 		return view('admin.posts.index', compact('posts'));
 	}
 	// public function create()
@@ -25,24 +28,31 @@ class PostsController extends Controller
 	// }
 	public function store(Request $request)
 	{
-		$this->validate($request, ['title' => 'required']);
-		$post = Post::create( $request->only('title') );
+		$this->authorize('create', new Post);
+		$this->validate($request, ['title' => 'required|min:3']);
+		// $post = Post::create( $request->only('title') );
+		$post = Post::create( $request->all() );
 		return redirect()->route('admin.posts.edit', $post);
 	}
 	public function edit(Post $post)
 	{
-		$categories = Category::all();
-		$tags 		= Tag::all();
-		return view('admin.posts.edit', compact('categories', 'tags', 'post'));
+		$this->authorize('update', $post);
+		return view('admin.posts.edit', [
+			'post'		 => $post,
+			'tags' 		 => Tag::all(),
+			'categories' => Category::all()
+		]);
 	}
 	public function update(Post $post, StorePostRequest $request)
 	{
+		$this->authorize('update', $post);
 		$post->update($request->all());
 		$post->syncTags($request->get('tags'));
 		return redirect()->route('admin.posts.edit', $post)->with('flash', 'La publicación ha sido guardada');
 	}
 	public function destroy(Post $post)
 	{
+		$this->authorize('delete', $post);
 		$post->delete();
 		return redirect()->route('admin.posts.index')->with('flash', 'La publicación ha sido eliminada');
 	}
